# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyApp",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/spectral-modes": {
      "get": {
        "tags": [
          "spectral-modes"
        ],
        "summary": "List Spectral Modes Endpoint",
        "description": "分页光谱模式.",
        "operationId": "list_spectral_modes_endpoint_api_v1_spectral_modes_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpectralModeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "spectral-modes"
        ],
        "summary": "Create Spectral Mode Endpoint",
        "description": "创建模式.",
        "operationId": "create_spectral_mode_endpoint_api_v1_spectral_modes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpectralModeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpectralModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spectral-modes/{mode_id}": {
      "get": {
        "tags": [
          "spectral-modes"
        ],
        "summary": "Get Spectral Mode Endpoint",
        "description": "获取模式.",
        "operationId": "get_spectral_mode_endpoint_api_v1_spectral_modes__mode_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpectralModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "spectral-modes"
        ],
        "summary": "Update Spectral Mode Endpoint",
        "description": "更新模式.",
        "operationId": "update_spectral_mode_endpoint_api_v1_spectral_modes__mode_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mode Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpectralModeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpectralModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "spectral-modes"
        ],
        "summary": "Delete Spectral Mode Endpoint",
        "description": "删除模式.",
        "operationId": "delete_spectral_mode_endpoint_api_v1_spectral_modes__mode_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mode Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/samples": {
      "get": {
        "tags": [
          "samples"
        ],
        "summary": "List Project Samples Endpoint",
        "description": "项目下样本列表.",
        "operationId": "list_project_samples_endpoint_api_v1_projects__project_id__samples_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/samples/{sample_id}": {
      "get": {
        "tags": [
          "samples"
        ],
        "summary": "Get Sample Detail Endpoint",
        "description": "样本详情.",
        "operationId": "get_sample_detail_endpoint_api_v1_samples__sample_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationSampleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "samples"
        ],
        "summary": "Update Sample Status Endpoint",
        "description": "更新样本状态.",
        "operationId": "update_sample_status_endpoint_api_v1_samples__sample_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sample Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationSampleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/samples/{sample_id}/annotations": {
      "put": {
        "tags": [
          "samples"
        ],
        "summary": "Replace Sample Annotations Endpoint",
        "description": "保存样本标注.",
        "operationId": "replace_sample_annotations_endpoint_api_v1_samples__sample_id__annotations_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sample Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleAnnotationsPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationSampleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/samples/{sample_id}/assets": {
      "get": {
        "tags": [
          "samples"
        ],
        "summary": "Get Sample Asset Endpoint",
        "description": "下载样本文件.",
        "operationId": "get_sample_asset_endpoint_api_v1_samples__sample_id__assets_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sample Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "样本相对路径",
              "title": "Path"
            },
            "description": "样本相对路径"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnnotationSampleDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "sample_id": {
            "type": "string",
            "title": "Sample Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "sample_type": {
            "type": "string",
            "title": "Sample Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_annotated": {
            "type": "boolean",
            "title": "Is Annotated"
          },
          "last_annotated_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Annotated By"
          },
          "source_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Files"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "annotations": {
            "items": {
              "$ref": "#/components/schemas/AnnotationDetailResponse"
            },
            "type": "array",
            "title": "Annotations"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sample_id",
          "project_id",
          "sample_type",
          "status",
          "is_annotated",
          "source_files",
          "created_at",
          "updated_at"
        ],
        "title": "AnnotationSampleDetail",
        "description": "样本详情."
      },
      "AnnotationSampleSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "sample_id": {
            "type": "string",
            "title": "Sample Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "sample_type": {
            "type": "string",
            "title": "Sample Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_annotated": {
            "type": "boolean",
            "title": "Is Annotated"
          },
          "last_annotated_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Annotated By"
          },
          "source_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Files"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "has_annotations": {
            "type": "boolean",
            "title": "Has Annotations"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sample_id",
          "project_id",
          "sample_type",
          "status",
          "is_annotated",
          "source_files",
          "created_at",
          "updated_at",
          "has_annotations"
        ],
        "title": "AnnotationSampleSummary",
        "description": "样本摘要."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SampleAnnotationsPayload": {
        "properties": {
          "annotations": {
            "items": {
              "$ref": "#/components/schemas/AnnotationDetailCreate"
            },
            "type": "array",
            "title": "Annotations"
          },
          "mark_annotated": {
            "type": "boolean",
            "title": "Mark Annotated",
            "default": true
          }
        },
        "type": "object",
        "title": "SampleAnnotationsPayload",
        "description": "保存标注载荷."
      },
      "SampleListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnnotationSampleSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SampleListResponse",
        "description": "样本列表."
      },
      "SampleStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "valid",
              "ignored"
            ],
            "title": "Status"
          },
          "is_annotated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Annotated"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SampleStatusUpdate",
        "description": "样本状态更新."
      },
      "SpectralModeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "r_channel": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "R Channel"
          },
          "g_channel": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "G Channel"
          },
          "b_channel": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "B Channel"
          },
          "r_gain": {
            "type": "number",
            "maximum": 4096.0,
            "minimum": -4096.0,
            "title": "R Gain",
            "default": 1.0
          },
          "g_gain": {
            "type": "number",
            "maximum": 4096.0,
            "minimum": -4096.0,
            "title": "G Gain",
            "default": 1.0
          },
          "b_gain": {
            "type": "number",
            "maximum": 4096.0,
            "minimum": -4096.0,
            "title": "B Gain",
            "default": 1.0
          },
          "gain_algorithm": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Gain Algorithm",
            "default": "linear"
          },
          "dark_calibration": {
            "type": "boolean",
            "title": "Dark Calibration",
            "default": false
          },
          "white_calibration": {
            "type": "boolean",
            "title": "White Calibration",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "r_channel",
          "g_channel",
          "b_channel"
        ],
        "title": "SpectralModeCreate",
        "description": "创建模式."
      },
      "SpectralModeListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SpectralModeResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "SpectralModeListResponse",
        "description": "分页响应."
      },
      "SpectralModeResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "r_channel": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "R Channel"
          },
          "g_channel": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "G Channel"
          },
          "b_channel": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "B Channel"
          },
          "r_gain": {
            "type": "number",
            "maximum": 4096.0,
            "minimum": -4096.0,
            "title": "R Gain",
            "default": 1.0
          },
          "g_gain": {
            "type": "number",
            "maximum": 4096.0,
            "minimum": -4096.0,
            "title": "G Gain",
            "default": 1.0
          },
          "b_gain": {
            "type": "number",
            "maximum": 4096.0,
            "minimum": -4096.0,
            "title": "B Gain",
            "default": 1.0
          },
          "gain_algorithm": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Gain Algorithm",
            "default": "linear"
          },
          "dark_calibration": {
            "type": "boolean",
            "title": "Dark Calibration",
            "default": false
          },
          "white_calibration": {
            "type": "boolean",
            "title": "White Calibration",
            "default": false
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "r_channel",
          "g_channel",
          "b_channel",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "SpectralModeResponse",
        "description": "响应."
      },
      "SpectralModeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "r_channel": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "R Channel"
          },
          "g_channel": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "G Channel"
          },
          "b_channel": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "B Channel"
          },
          "r_gain": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 4096.0,
                "minimum": -4096.0
              },
              {
                "type": "null"
              }
            ],
            "title": "R Gain"
          },
          "g_gain": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 4096.0,
                "minimum": -4096.0
              },
              {
                "type": "null"
              }
            ],
            "title": "G Gain"
          },
          "b_gain": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 4096.0,
                "minimum": -4096.0
              },
              {
                "type": "null"
              }
            ],
            "title": "B Gain"
          },
          "gain_algorithm": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Gain Algorithm"
          },
          "dark_calibration": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dark Calibration"
          },
          "white_calibration": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "White Calibration"
          }
        },
        "type": "object",
        "title": "SpectralModeUpdate",
        "description": "更新模式."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}